+++ src/main/java/lotto/domain/LottoGame.java
@@ -1,0 +1,10 @@
+package lotto.domain;
+
+import lotto.domain.lotto.Lottos;
+
+public class LottoGame {
+
+    private Lottos lottos;
+
+
+}
