+++ src/main/java/lotto/domain/lotto/Lottos.java
@@ -1,15 +1,15 @@
-package lotto.domain.wrapper;
+package lotto.domain.lotto;
 
-import lotto.domain.LotteryRank;
-import lotto.domain.Lotto;
-import lotto.domain.RankCount;
+import lotto.domain.rankcount.RankCount;
+import lotto.domain.rankcount.RankCountGroup;
+import lotto.domain.wrapper.Numbers;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import static lotto.domain.LotteryRank.*;
+import static lotto.domain.lotto.LotteryRank.*;
 
 public class Lottos {
 
@@ -19,7 +19,7 @@ public class Lottos {
         numberses.forEach(numbers -> lottos.add(new Lotto(numbers)));
     }
 
-    public RankCountGroup groupByRank(Numbers winningNumbers) {
+    public RankCountGroup groupByRankCount(Numbers winningNumbers) {
         LotteryRank[] ranks = values();
 
         List<RankCount> rankCounts = Arrays.stream(ranks)
