+++ src/test/java/lotto/domain/LottoGameTest.java
@@ -1,0 +1,20 @@
+package lotto.domain;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class LottoGameTest {
+
+    @DisplayName("구입 금액과 총 당첨금을 나눠 수익률을 계산한다.")
+    @Test
+    void calculate() {
+        // given
+
+        // when
+
+        // then
+
+    }
+}
