+++ src/test/java/lotto/domain/lotto/LottoTest.java
@@ -1,7 +1,7 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
+import lotto.domain.lotto.Lotto;
 import lotto.domain.wrapper.Numbers;
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
